+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## master
+
+* Split library into `Console.BnbStaking` module hierarchy.
+
 ## v0.1.0.0
 
 * Initial release